feature: Add channel synchronization status
Instead of raising errors, we make sure to avoid that and store the
failed status in the channel itself, in order to provide a runtime
handling of the issues.

We store the status for the feed, new, success or failure and also the
time for the last synchronized time, which may comes handy in the near
future.

Besides all that, we also store the reason for the failure (the
last one at least).

Follwing the principle: Shit happens.