fix: Prevent failures for the channel for a single episode
Persist other episodes and continue channel-processing by ignoring
failed episodes.

Mongo handles the object parameters as embedded resources, what causes
the channel to have several episodes included (sometimes not-valid ones)
and that generates an error when we try to persist the failure reasons
on the rescue code.

e.g.:

  channel = Channel.find(...)
  episode = Episode.new(foo: bar, channel: channel, invalid: params)
  channel.save! -> raises error of invalid episode

The solution is to use channel_id to point the relation instead of the
channel object itself.

Not the perfect solution, however I am still analysing what is the best
option for the mongo "relations", embedded vs relational.