refactor: Extract Episodes synchronization loop to an operation
Now we have a more OO approach, better delegation of responsabilities
and such.

The Episode::SyncAll knows about the logic of when an episode fails and
handles it. While it clears Channel::Sync from that, it also makes it
easy to evolve the code with specifics of each `part` of the
synchronization pipeline.