refactor: Move episode url to an embedded audio object - Step 1
We need to evolve our database model in order to support new features.

The next big feature is analysing the episodes audio and collect
meta-data like duration, bitrate, samplerate, codec, and so ...

For that to "scale" we need a Audio object that only knows about audio
properties inside our episode's file.

Why Step 1?

Mongoid is schemaless, soon we need to keep the reference for the url
field before moving it to audio, the step 2 will be to completetly
remove the references, hopefully, in the next commit.