feature: Add audio analysis pipeline
Now, every audio has an status and it can be "analysed",^ "failed" or
"new".

The Analysis consists on downloading the file, running a ffmpeg to check
its own parameters, like duration, size, codec, bitrate and sample_rate
... then after we store those infos in the audio instance.

Heroku is going to run the make analyse partially the audios, every N minutes.

In an ideal world we would use a pub-sub and create a pipeline where
after publication the episode would be analysed, but we have
infrastructure limitations.

The public interface for running right now is: make analyse
Which will call synchronize for the last 25 episodes.