fix: Prevent it from getting stuck on failing audios
Currently, we select the last N audios, based on publication date and
then we start to analyse them. However, if they fail for some reason, we
try again in the next run. That works well for recent audios, but after
N failing audios we get stuck on those N audios, because they are the
last N non-analysed ones.

In order to prevent that we are going to use a erro counter, an
episode is declared officially failed after 3 attempts, then it will not
be taken in consideration for the next runs.