refactor: Use mongo query instead of memory find
It seems that the previous way we were loading all the episodes in
memory to find the correct one, now we can force to use only the
expected one.

This is still not a good way of doing it, however, it is better then the
current.

The best way would be to use a cached collection, I will work on that
soon enough!