feature: Store info of an user listening an episode
It will be useful to have information about our users, when do they
access what.

In order to make it possible to track them now, we set a permanent
cookie, to make sure we have an "ID".

From that ID we create a "episode-listen" entry, which represent that a
user has accessed an episode and when.

In the future that could inform how far did the user listens to, did he
finish? ...