refactor: Use sidekiq instead of crontasks
It is about time that we have a dedicated worker with threads running to
split the download/process of the tasks. Also, we can increase the
amount of processing.

For the next things, like, data pipeline, it will be really useful to
have the worker managing the exports and triggering of the APIs for
recommendation content.